+++ server/src/test/java/com/adobe/testing/s3mock/its/AmazonClientUploadIT.java
@@ -896,138 +896,6 @@ class AmazonClientUploadIT extends S3TestBase {
   }
 
   /**
-   * Tests if user metadata can be passed by multipart upload.
-   */
-  @Test
-  void shouldPassUserMetadataWithMultipartUploads() {
-    s3Client.createBucket(BUCKET_NAME);
-
-    final File uploadFile = new File(UPLOAD_FILE_NAME);
-    final ObjectMetadata objectMetadata = new ObjectMetadata();
-    objectMetadata.addUserMetadata("key", "value");
-
-    final InitiateMultipartUploadResult initiateMultipartUploadResult = s3Client
-        .initiateMultipartUpload(
-            new InitiateMultipartUploadRequest(BUCKET_NAME, UPLOAD_FILE_NAME, objectMetadata));
-    final String uploadId = initiateMultipartUploadResult.getUploadId();
-
-    final UploadPartResult uploadPartResult = s3Client.uploadPart(new UploadPartRequest()
-        .withBucketName(initiateMultipartUploadResult.getBucketName())
-        .withKey(initiateMultipartUploadResult.getKey())
-        .withUploadId(uploadId)
-        .withFile(uploadFile)
-        .withFileOffset(0)
-        .withPartNumber(1)
-        .withPartSize(uploadFile.length())
-        .withLastPart(true));
-
-    final List<PartETag> partETags = singletonList(uploadPartResult.getPartETag());
-    s3Client.completeMultipartUpload(new CompleteMultipartUploadRequest(
-        initiateMultipartUploadResult.getBucketName(),
-        initiateMultipartUploadResult.getKey(),
-        initiateMultipartUploadResult.getUploadId(),
-        partETags
-    ));
-
-    final ObjectMetadata metadataExisting = s3Client.getObjectMetadata(
-        initiateMultipartUploadResult.getBucketName(), initiateMultipartUploadResult.getKey()
-    );
-
-    assertThat("User metadata should be identical!", metadataExisting.getUserMetadata(),
-        is(equalTo(objectMetadata.getUserMetadata())));
-  }
-
-  @Test
-  void shouldInitiateMultipartAndRetrieveParts() throws IOException {
-    s3Client.createBucket(BUCKET_NAME);
-
-    final File uploadFile = new File(UPLOAD_FILE_NAME);
-    final ObjectMetadata objectMetadata = new ObjectMetadata();
-    final String hash = DigestUtils.md5Hex(new FileInputStream(uploadFile));
-    final String expectedEtag = String.format("%s-%s", hash, 1);
-    objectMetadata.addUserMetadata("key", "value");
-
-    final InitiateMultipartUploadResult initiateMultipartUploadResult = s3Client
-        .initiateMultipartUpload(
-            new InitiateMultipartUploadRequest(BUCKET_NAME, UPLOAD_FILE_NAME, objectMetadata));
-    final String uploadId = initiateMultipartUploadResult.getUploadId();
-    final String key = initiateMultipartUploadResult.getKey();
-
-    s3Client.uploadPart(new UploadPartRequest()
-        .withBucketName(initiateMultipartUploadResult.getBucketName())
-        .withKey(initiateMultipartUploadResult.getKey())
-        .withUploadId(uploadId)
-        .withFile(uploadFile)
-        .withFileOffset(0)
-        .withPartNumber(1)
-        .withPartSize(uploadFile.length())
-        .withLastPart(true));
-
-    final ListPartsRequest listPartsRequest = new ListPartsRequest(
-        BUCKET_NAME,
-        key,
-        uploadId
-    );
-    final PartListing partListing = s3Client.listParts(listPartsRequest);
-
-    assertThat("Part listing should be 1", partListing.getParts().size(), is(1));
-    final PartSummary partSummary = partListing.getParts().get(0);
-
-    assertThat("Etag should match", partSummary.getETag(), is(expectedEtag));
-    assertThat("Part number should match", partSummary.getPartNumber(), is(1));
-    assertThat("LastModified should be valid date", partSummary.getLastModified(), any(Date.class));
-
-  }
-
-  /**
-   * Tests if not yet completed / aborted multipart uploads are listed.
-   */
-  @Test
-  void shouldListMultipartUploads() {
-    s3Client.createBucket(BUCKET_NAME);
-
-    assertThat(s3Client.listMultipartUploads(new ListMultipartUploadsRequest(BUCKET_NAME))
-        .getMultipartUploads(), is(empty()));
-
-    final InitiateMultipartUploadResult initiateMultipartUploadResult = s3Client
-        .initiateMultipartUpload(new InitiateMultipartUploadRequest(BUCKET_NAME, UPLOAD_FILE_NAME));
-    final String uploadId = initiateMultipartUploadResult.getUploadId();
-
-    final MultipartUploadListing listing =
-        s3Client.listMultipartUploads(new ListMultipartUploadsRequest(BUCKET_NAME));
-    assertThat(listing.getMultipartUploads(), is(not(empty())));
-    assertThat(listing.getBucketName(), equalTo(BUCKET_NAME));
-    assertThat(listing.getMultipartUploads(), hasSize(1));
-    final MultipartUpload upload = listing.getMultipartUploads().get(0);
-    assertThat(upload.getUploadId(), equalTo(uploadId));
-    assertThat(upload.getKey(), equalTo(UPLOAD_FILE_NAME));
-  }
-
-  /**
-   * Tests if a multipart upload can be aborted.
-   */
-  @Test
-  void shouldAbortMultipartUpload() {
-    s3Client.createBucket(BUCKET_NAME);
-
-    assertThat(s3Client.listMultipartUploads(new ListMultipartUploadsRequest(BUCKET_NAME))
-        .getMultipartUploads(), is(empty()));
-
-    final InitiateMultipartUploadResult initiateMultipartUploadResult = s3Client
-        .initiateMultipartUpload(new InitiateMultipartUploadRequest(BUCKET_NAME, UPLOAD_FILE_NAME));
-    final String uploadId = initiateMultipartUploadResult.getUploadId();
-
-    assertThat(s3Client.listMultipartUploads(new ListMultipartUploadsRequest(BUCKET_NAME))
-        .getMultipartUploads(), is(not(empty())));
-
-    s3Client.abortMultipartUpload(
-        new AbortMultipartUploadRequest(BUCKET_NAME, UPLOAD_FILE_NAME, uploadId));
-
-    assertThat(s3Client.listMultipartUploads(new ListMultipartUploadsRequest(BUCKET_NAME))
-        .getMultipartUploads(), is(empty()));
-  }
-
-  /**
    * Verify that range-downloads work.
    *
    * @throws Exception not expected
